!2284 导致版块排序设置bug * simple字段二进制存储字段，因无法用 tinyint(1) signed，故此改为 smallint

+++ upload/install/data/install.sql
@@ -2421,7 +2421,7 @@ CREATE TABLE pre_forum_forum (
   catforumcolumns tinyint(3) unsigned NOT NULL DEFAULT '0',
   threadcaches tinyint(1) NOT NULL DEFAULT '0',
   alloweditpost tinyint(1) NOT NULL DEFAULT '1',
-  `simple` tinyint(1) NOT NULL DEFAULT '0',
+  `simple` smallint(6) NOT NULL DEFAULT '0',
   modworks tinyint(1) NOT NULL DEFAULT '0',
   allowglobalstick tinyint(1) NOT NULL DEFAULT '1',
   `level` smallint(6) NOT NULL DEFAULT '0',
